+++ toucan_connectors/mssql/mssql_connector.py
@@ -1,18 +1,24 @@
-from contextlib import suppress
 from logging import getLogger
-from typing import Annotated
+from typing import TYPE_CHECKING, Annotated
 
-from pydantic import Field, StringConstraints, create_model
+from pydantic import Field, StringConstraints, create_model, model_validator
 
 try:
-    import pyodbc
+    import pandas as pd
+    from sqlalchemy.orm import Session
 
     CONNECTOR_OK = True
 except ImportError as exc:  # pragma: no cover
     getLogger(__name__).warning(f"Missing dependencies for {__name__}: {exc}")
     CONNECTOR_OK = False
 
-from toucan_connectors.common import pandas_read_sql
+from toucan_connectors.common import (
+    convert_jinja_params_to_sqlalchemy_named,
+    create_sqlalchemy_engine,
+    pandas_read_sqlalchemy_query,
+    pyformat_params_to_jinja,
+    unnest_sql_jinja_parameters,
+)
 from toucan_connectors.toucan_connector import (
     PlainJsonSecretStr,
     ToucanConnector,
@@ -20,35 +26,37 @@ from toucan_connectors.toucan_connector import (
     strlist_to_enum,
 )
 
+if TYPE_CHECKING:
+    import sqlalchemy as sa
+
 
 class MSSQLDataSource(ToucanDataSource):
     # By default SQL Server selects the database which is set
     # as default for specific user
-    database: str = Field(
+    database: str | None = Field(
         None,
         description="The name of the database you want to query. "
         "By default SQL Server selects the user's default database",
     )
-    table: Annotated[str, StringConstraints(min_length=1)] = Field(
+    table: Annotated[str, StringConstraints(min_length=1)] | None = Field(
         None,
         description='The name of the data table that you want to get (equivalent to "SELECT * FROM your_table")',
     )
-    query: Annotated[str, StringConstraints(min_length=1)] = Field(
+    query: Annotated[str, StringConstraints(min_length=1)] | None = Field(
         None,
         description="You can write a custom query against your "
         "database here. It will take precedence over "
         'the "table" parameter above',
-        widget="sql",
+        json_schema_extra={"widget": "sql"},
     )
 
-    def __init__(self, **data):
-        super().__init__(**data)
-        query = data.get("query")
-        table = data.get("table")
-        if query is None and table is None:
+    @model_validator(mode="after")
+    def table_or_query(self):
+        if self.query is None and self.table is None:
             raise ValueError("'query' or 'table' must be set")
-        elif query is None and table is not None:
-            self.query = f"select * from {table};"
+        elif self.query is None and self.table is not None:
+            self.query = f"select * from {self.table};"
+        return self
 
     @classmethod
     def get_form(cls, connector: "MSSQLConnector", current_config):
@@ -60,24 +68,27 @@ class MSSQLDataSource(ToucanDataSource):
         """
         constraints = {}
 
-        with suppress(Exception):
-            connection = pyodbc.connect(
-                **connector.get_connection_params(database=current_config.get("database", "tempdb"))
-            )
-            # # Always add the suggestions for the available databases
-            with connection.cursor() as cursor:
-                cursor.execute("SELECT name FROM sys.databases;")
+        sa_engine = connector._create_engine(database=current_config.get("database", "tempdb"))
+
+        # Always add the suggestions for the available databases
+        with Session(sa_engine) as session:
+            with session.connection() as connection:
+                cursor = connection.connection.cursor()
+                cursor.execute("SELECT name FROM sys.databases")
                 res = cursor.fetchall()
                 available_dbs = [r[0] for r in res]
+
                 constraints["database"] = strlist_to_enum("database", available_dbs)
 
                 if "database" in current_config:
-                    cursor.execute("SELECT TABLE_NAME FROM  INFORMATION_SCHEMA.TABLES;")
+                    cursor.execute("SELECT TABLE_NAME FROM INFORMATION_SCHEMA.TABLES;")
                     res = cursor.fetchall()
                     available_tables = [table_name for (table_name,) in res]
                     constraints["table"] = strlist_to_enum("table", available_tables, None)
 
-        return create_model("FormSchema", **constraints, __base__=cls).schema()
+                cursor.close()
+
+        return create_model("FormSchema", **constraints, __base__=cls).schema()  # type:ignore[call-overload]
 
 
 class MSSQLConnector(ToucanConnector, data_source_model=MSSQLDataSource):
@@ -91,45 +102,59 @@ class MSSQLConnector(ToucanConnector, data_source_model=MSSQLDataSource):
         "the hardcoded IP address of your database server",
     )
 
-    port: int = Field(None, description="The listening port of your database server")
+    port: int | None = Field(None, description="The listening port of your database server")
     user: str = Field(..., description="Your login username")
-    password: PlainJsonSecretStr = Field(None, description="Your login password")
-    connect_timeout: int = Field(
+    password: PlainJsonSecretStr | None = Field(None, description="Your login password")
+    connect_timeout: int | None = Field(
         None,
         title="Connection timeout",
         description="You can set a connection timeout in seconds here, i.e. the maximum length "
         "of time you want to wait for the server to respond. None by default",
     )
+    trust_server_certificate: bool = Field(
+        False,
+        title="Trust server certificate",
+        description="This allows to disable server certificate validation, which can be "
+        "required for custom and self-signed certificates. Connection is still encrypted.",
+    )
 
-    def get_connection_params(self, database):
+    def _create_engine(self, database: str | None) -> "sa.Engine":
+        from sqlalchemy.engine import URL
+
         server = self.host
         if server == "localhost":
             server = "127.0.0.1"  # localhost is not understood by pyodbc
         if self.port is not None:
             server += f",{self.port}"
-        con_params = {
-            "driver": "{ODBC Driver 18 for SQL Server}",
-            "server": server,
-            "database": database,
-            "user": self.user,
-            "password": self.password.get_secret_value() if self.password else None,
-            "timeout": self.connect_timeout,
-            "as_dict": True,
-        }
-        # remove None values
-        return {k: v for k, v in con_params.items() if v is not None}
 
-    def _retrieve_data(self, datasource):
-        connection = pyodbc.connect(**self.get_connection_params(datasource.database))
+        query_params: dict[str, str] = {"driver": "ODBC Driver 18 for SQL Server"}
+        if self.connect_timeout:
+            query_params["timeout"] = str(self.connect_timeout)
+        if self.trust_server_certificate:
+            query_params["TrustServerCertificate"] = "yes"
 
-        query_params = datasource.parameters or {}
-        df = pandas_read_sql(
-            datasource.query,
-            con=connection,
-            params=query_params,
-            convert_to_qmark=True,
-            render_user=True,
+        connection_url = URL.create(
+            "mssql+pyodbc",
+            username=self.user,
+            password=self.password.get_secret_value() if self.password else None,
+            host=server,
+            database=database,
+            query=query_params,
         )
+        return create_sqlalchemy_engine(connection_url)
 
-        connection.close()
+    def _retrieve_data(self, datasource: MSSQLDataSource) -> "pd.DataFrame":
+        sa_engine = self._create_engine(database=datasource.database)
+
+        # This should not happen as it is checked by the data sources' model validator
+        if datasource.query is None:
+            raise ValueError("'query' or 'table' must be set")
+
+        # %()s -> {{}}
+        jinja_query = pyformat_params_to_jinja(datasource.query)
+        flattened_query, flattened_params = unnest_sql_jinja_parameters(jinja_query, datasource.parameters or {})
+        final_query = convert_jinja_params_to_sqlalchemy_named(flattened_query)
+
+        df = pandas_read_sqlalchemy_query(query=final_query, engine=sa_engine, params=flattened_params)
+
         return df